+++ web/src/main/webapp/web/js/views/common/view.js
@@ -31,11 +31,13 @@ define([
             );
         },
         delete: function(event) {
-            this.model.destroy({
-                wait: true,
-                success: NotificationHandler.onModelDeleteSuccess,
-                error: NotificationHandler.onServerError
-            });
+            NotificationHandler.confirmAction(function() {
+                this.model.destroy({
+                    wait: true,
+                    success: NotificationHandler.onModelDeleteSuccess,
+                    error: NotificationHandler.onServerError
+                });
+            }, this);
         },
         close: function() {
             this.unbind();
