+++ web/src/main/webapp/web/js/views/participants/view.js
@@ -123,11 +123,13 @@ define([
             }
         },
         removeParticipant: function() {
-            this.model.destroy({
-                wait: true,
-                success: NotificationHandler.onModelDeleteSuccess,
-                error: NotificationHandler.onServerError
-            });
+            NotificationHandler.confirmAction(function() {
+                this.model.destroy({
+                    wait: true,
+                    success: NotificationHandler.onModelDeleteSuccess,
+                    error: NotificationHandler.onServerError
+                });
+            }, this);
         },
         close: function() {
             this.viewHolder.closeAll();
